// frontend2/src/api/orval/client/model/canvasShapeObjectOutput.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Military Journal API
 * OpenAPI spec version: 1.0.0
 */
import type { CanvasObjectProperties } from './canvasObjectProperties';
import type { CanvasObjectStyle } from './canvasObjectStyle';
import type { CanvasObjectTransform } from './canvasObjectTransform';
import type { ShapeField } from './shapeField';

export interface CanvasShapeObjectOutput {
  id: string;
  type: 'shape';
  transform: CanvasObjectTransform;
  style?: CanvasObjectStyle | null;
  properties?: CanvasObjectProperties | null;
  fields: ShapeField;
}