// frontend2/src/api/orval/client/model/canvasContentInput.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Military Journal API
 * OpenAPI spec version: 1.0.0
 */
import type { CanvasContentInputMetadata } from './canvasContentInputMetadata';
import type { CanvasImageObject } from './canvasImageObject';
import type { CanvasShapeObjectInput } from './canvasShapeObjectInput';
import type { CanvasSymbolObject } from './canvasSymbolObject';
import type { CanvasTextObject } from './canvasTextObject';

export interface CanvasContentInput {
  version?: string;
  canvasWidth?: number;
  canvasHeight?: number;
  backgroundColor?: string;
  objects?: (CanvasTextObject | CanvasImageObject | CanvasShapeObjectInput | CanvasSymbolObject)[];
  metadata?: CanvasContentInputMetadata;
}